pyproject.toml

[project]
name = "industrial_safety_project"
version = "1.0.0"
description = ""
authors = [
    {name = "vasya", email = "spam@shit.com"},
]
dependencies = [
    "django>=4.2.1",
    "djangorestframework>=3.14.0",
    "pyyaml>=6.0",
    "uritemplate>=4.1.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}
